requirements.txt
flask
opentelemetry-api
opentelemetry-sdk
opentelemetry-instrumentation-flask
opentelemetry-exporter-jaeger
opentelemetry-exporter-otlp
opentelemetry-exporter-prometheus
prometheus-flask-exporter